requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==3.2.12
django-model-utils==4.2.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
psycopg2==2.9.5
psycopg2-binary==2.9.5
PyJWT==2.6.0
pytz==2022.6
sqlparse==0.4.3
gunicorn==20.1.0
